refactor: 更好的component类型
